+++ mod/pcoder/mod/php-diff-1.0/lib/Diff/Renderer/Html/Inline.php
@@ -61,9 +61,9 @@ class Diff_Renderer_Html_Inline extends Diff_Renderer_Html_Array
 		$html .= '<table class="Differences DifferencesInline">';
 		$html .= '<thead>';
 		$html .= '<tr>';
-		$html .= '<th>Old</th>';
-		$html .= '<th>New</th>';
-		$html .= '<th>Differences</th>';
+		$html .= '<th>#1</th>';
+		$html .= '<th>#2</th>';
+		$html .= '<th>Diferencias / Differences</th>';
 		$html .= '</tr>';
 		$html .= '</thead>';
 		foreach($changes as $i => $blocks) {